refactor(built-in-mcp): clarify plugin ownership and credential boundaries
