+++ src/backend/services/builtInMcp/authorization/PluginCredentialStore.ts
@@ -1,22 +1,23 @@
 import { randomUUID } from 'expo-crypto';
 import * as SecureStore from 'expo-secure-store';
 
+import { PluginSecretReferenceSchema, type PluginSecretReference } from '@/backend/data/db/schemas';
 import {
   pluginAuthorizationService,
   type PluginAuthorizationService,
 } from '@/backend/data/services/PluginAuthorizationService';
-import {
-  PluginSecretReferenceSchema,
-  type PluginSecretReference,
-} from '@/backend/data/services/utils/pluginSecretReferences';
 import { PluginError } from '@/shared/contracts/plugins';
-import { PluginCredentialSchema, type PluginCredential } from '@/shared/data/types/plugin';
 
-import type { PluginAuthorizationStore } from './pluginDefinition';
+import type { PluginAuthorizationStore } from './pluginAuthorization';
+import {
+  PluginCredentialSchema,
+  type PluginCredential,
+  type PluginGrant,
+} from './pluginCredential';
 
 type Database = Pick<
   PluginAuthorizationService,
-  'connect' | 'disconnect' | 'getCredentialGrant' | 'getCurrentGrant'
+  'connect' | 'disconnect' | 'getAuthorizedGrant' | 'getCurrentGrant'
 >;
 type NativeStorage = Pick<typeof SecureStore, 'getItemAsync' | 'setItemAsync' | 'deleteItemAsync'>;
 const OPTIONS = { keychainAccessible: SecureStore.WHEN_UNLOCKED_THIS_DEVICE_ONLY };
@@ -69,24 +70,24 @@ export class PluginCredentialStore {
     );
   }
 
-  private async resolveCredential(reference: PluginCredential) {
+  private async resolveCredential(reference: PluginSecretReference) {
     const value = await this.read(secretKey(PluginSecretReferenceSchema.parse(reference)));
     if (!value)
       throw new PluginError('authorization', 'Plugin credentials are missing. Connect again.');
     return value;
   }
 
-  private async remove(reference: PluginCredential) {
+  private async remove(reference: PluginSecretReference) {
     // Cleanup cannot undo a committed replacement or revocation. No background retries.
     await this.native
       .deleteItemAsync(secretKey(PluginSecretReferenceSchema.parse(reference)), OPTIONS)
       .catch(() => {});
   }
 
-  getCredentialGrant(pluginId: string, authorizationId: string) {
+  getCredentialGrant(pluginId: string, authorizationId: string): Promise<PluginGrant> {
     return this.run(async () => {
-      const row = await this.database.getCredentialGrant(pluginId, authorizationId);
-      return { ...row, credential: await this.resolveCredential(row.credential) };
+      const row = await this.database.getAuthorizedGrant(pluginId, authorizationId);
+      return { id: row.id, credential: await this.resolveCredential(row.credentialReference) };
     });
   }
 
@@ -108,7 +109,7 @@ export class PluginCredentialStore {
         this.run(async () => {
           const row = await this.database.getCurrentGrant(pluginId, method);
           if (!row || (id && row.id !== id)) return undefined;
-          return { id: row.id, credential: await this.resolveCredential(row.credential) };
+          return { id: row.id, credential: await this.resolveCredential(row.credentialReference) };
         }),
       updateCredential: (id, credential, signal) =>
         this.run(async () => {
@@ -118,7 +119,7 @@ export class PluginCredentialStore {
           signal.throwIfAborted();
           // Replace the entire token bundle in one native item; no second SQLite write.
           await this.write(
-            secretKey(PluginSecretReferenceSchema.parse(row.credential)),
+            secretKey(PluginSecretReferenceSchema.parse(row.credentialReference)),
             credential,
           );
           return true;
@@ -137,7 +138,12 @@ export class PluginCredentialStore {
     };
   }
 
-  connect(input: Parameters<PluginAuthorizationService['connect']>[0], callerSignal?: AbortSignal) {
+  connect(
+    input: Omit<Parameters<PluginAuthorizationService['connect']>[0], 'credentialReference'> & {
+      credential: PluginCredential;
+    },
+    callerSignal?: AbortSignal,
+  ) {
     const signal = callerSignal
       ? AbortSignal.any([callerSignal, this.lifetime.signal])
       : this.lifetime.signal;
@@ -145,15 +151,19 @@ export class PluginCredentialStore {
       signal.throwIfAborted();
       const previous = await this.database.getCurrentGrant(input.pluginId);
       const reference = { storage: 'secure-store-v1' as const, id: randomUUID() };
+      const { credential, ...connectionInput } = input;
       let connection;
       try {
-        await this.write(secretKey(reference), input.credential);
-        connection = await this.database.connect({ ...input, credential: reference }, signal);
+        await this.write(secretKey(reference), credential);
+        connection = await this.database.connect(
+          { ...connectionInput, credentialReference: reference },
+          signal,
+        );
       } catch (error) {
         await this.remove(reference);
         throw error;
       }
-      if (previous) await this.remove(previous.credential);
+      if (previous) await this.remove(previous.credentialReference);
       return connection;
     });
   }
@@ -162,7 +172,7 @@ export class PluginCredentialStore {
     return this.run(async () => {
       const previous = await this.database.getCurrentGrant(pluginId);
       const result = await this.database.disconnect(pluginId);
-      if (previous) await this.remove(previous.credential);
+      if (previous) await this.remove(previous.credentialReference);
       return result;
     });
   }
