+++ src/backend/services/builtInMcp/authorization/__tests__/PluginCredentialStore.test.ts
