+++ src/backend/services/builtInMcp/index.ts
@@ -1,4 +1,4 @@
-export { createBuiltInMcpClient, isBuiltInMcpToolAllowed } from './createBuiltInMcpClient';
+export { createBuiltInMcpClient } from './transport/createBuiltInMcpClient';
 export { createPluginsModule } from './createPluginsModule';
-export { getBuiltInPluginCatalog } from './pluginRegistry';
-export { PluginAuthorizationManager } from './PluginAuthorizationManager';
+export { getBuiltInPluginCatalog, isBuiltInMcpToolAllowed } from './pluginRegistry';
+export { PluginAuthorizationManager } from './authorization/PluginAuthorizationManager';
