+++ src/backend/services/builtInMcp/plugins/feishu/index.ts
@@ -1,0 +1,1 @@
+export { feishuPlugin } from './feishuPlugin';
