+++ src/frontend/features/plugin/detail/connect/InteractiveConnect.tsx
@@ -1,31 +1,17 @@
-import { Button, ContentState, useAlert, useToast } from '@cherrystudio/ui/components';
-import { useFocusEffect, useRouter } from 'expo-router';
-import * as WebBrowser from 'expo-web-browser';
-import { useCallback, useEffect, useEffectEvent, useRef, useState, type ReactNode } from 'react';
+import { Button, ContentState, useAlert } from '@cherrystudio/ui/components';
+import type { ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
-import { AppState, Keyboard, Linking, Text, View } from 'react-native';
+import { Text, View } from 'react-native';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-controller';
 
 import { RouteHeader } from '@/frontend/appShell/header';
-import { useBackendModule } from '@/frontend/data';
 import { keyboardBottomOffset } from '@/frontend/utils/constants';
-import {
-  PluginError,
-  type PluginAuthorizationObservation,
-  type PluginAuthorizationState,
-  type PluginErrorReason,
-} from '@/shared/contracts/plugins';
 import type { PluginCatalogEntry, PluginInteractiveMethod } from '@/shared/data/types/plugin';
-import { createPluginCredentialsSchema } from '@/shared/utils/pluginCredentials';
 
-import { useRefreshPluginConnections } from '../../usePluginConnections';
 import { CredentialFields, hasEveryField } from './CredentialFields';
+import { useInteractiveConnect } from './useInteractiveConnect';
 
-/**
- * Renders the backend's authorization observation and forwards user actions. Polling,
- * completion and retry bookkeeping live in the backend observer, attached only while this
- * route is focused and the app is active.
- */
+/** Presents browser-confirmation stages and the optional existing-application form. */
 export function InteractiveConnect({
   entry,
   method,
@@ -35,120 +21,24 @@ export function InteractiveConnect({
   method: PluginInteractiveMethod;
   children: ReactNode;
 }) {
-  const plugins = useBackendModule('plugins');
   const { t } = useTranslation();
-  const router = useRouter();
-  const { toast } = useToast();
   const { alert } = useAlert();
-  const refresh = useRefreshPluginConnections();
-  const [observation, setObservation] = useState<PluginAuthorizationObservation | null>(null);
-  const [actionError, setActionError] = useState<PluginErrorReason | null>(null);
-  const [isActing, setIsActing] = useState(false);
-  const [existingApplication, setExistingApplication] = useState<{
-    fields: Record<string, string>;
-    invalid: Set<string>;
-  } | null>(null);
-  const finished = useRef(false);
+  const {
+    state,
+    isBusy,
+    error,
+    existingApplication,
+    setExistingApplication,
+    begin,
+    submitExistingApplication,
+    openConfirmation,
+    check,
+    cancel,
+    resetApplication,
+  } = useInteractiveConnect(entry, method);
   const name = t(`plugins.catalog.${entry.id}.name`);
   const textKey = `plugins.catalog.${entry.id}.authMethods.${method.id}`;
   const applicationFields = method.applicationFields;
-
-  useFocusEffect(
-    useCallback(() => {
-      let detach: (() => void) | null = null;
-      const attach = () => {
-        detach ??= plugins.authorization.observe(entry.id, method.id, setObservation);
-      };
-      const release = () => {
-        detach?.();
-        detach = null;
-      };
-      if (AppState.currentState === 'active') attach();
-      const listener = AppState.addEventListener('change', (status) =>
-        status === 'active' ? attach() : release(),
-      );
-      return () => {
-        listener.remove();
-        release();
-      };
-    }, [plugins, entry.id, method.id]),
-  );
-
-  const connection = observation?.connection;
-  const connected = useEffectEvent(async () => {
-    // Only an owned in-app presentation can be dismissed; external browsers return manually.
-    await WebBrowser.dismissBrowser().catch(() => undefined);
-    await refresh();
-    toast.show({ label: t('plugins.connectSuccess', { name }), variant: 'success' });
-    router.back();
-  });
-  useEffect(() => {
-    if (!connection || finished.current) return;
-    finished.current = true;
-    void connected();
-  }, [connection]);
-
-  async function openConfirmation(state: PluginAuthorizationState) {
-    if (state.status !== 'waiting') return;
-    try {
-      // Android may resolve immediately; iOS resolves on close (including `cancel`
-      // after successful approval). Neither result is proof of success or denial.
-      await WebBrowser.openBrowserAsync(state.verificationUrl).catch(() =>
-        Linking.openURL(state.verificationUrl),
-      );
-    } catch {
-      toast.show({ label: t('plugins.authorization.browserFailed'), variant: 'danger' });
-    } finally {
-      plugins.authorization.check(entry.id, method.id);
-    }
-  }
-
-  async function act(action: () => Promise<PluginAuthorizationState | void>) {
-    if (isActing) return;
-    setIsActing(true);
-    setActionError(null);
-    try {
-      return await action();
-    } catch (error) {
-      setActionError(error instanceof PluginError ? error.reason : 'request');
-      return undefined;
-    } finally {
-      setIsActing(false);
-    }
-  }
-
-  const begin = (restart = false) =>
-    act(async () => {
-      if (restart) await plugins.authorization.cancel(entry.id, method.id);
-      const next = await plugins.authorization.begin(entry.id, method.id);
-      void openConfirmation(next);
-      return next;
-    });
-
-  const submitExistingApplication = () =>
-    act(async () => {
-      if (!existingApplication || !applicationFields) return;
-      const parsed = createPluginCredentialsSchema(applicationFields).safeParse(
-        existingApplication.fields,
-      );
-      if (!parsed.success) {
-        setExistingApplication({
-          ...existingApplication,
-          invalid: new Set(parsed.error.issues.map((issue) => String(issue.path[0]))),
-        });
-        return;
-      }
-      Keyboard.dismiss();
-      // Keep credentials out of route parameters and query caches.
-      await plugins.authorization.useApplication(entry.id, method.id, parsed.data);
-      setExistingApplication(null);
-      const next = await plugins.authorization.begin(entry.id, method.id);
-      void openConfirmation(next);
-    });
-
-  const state = observation?.state ?? null;
-  const busy = isActing || observation?.busy === true;
-  const error = actionError ?? observation?.error ?? null;
   const waiting = state?.status === 'waiting' ? state : null;
   const finalStatus =
     state?.status === 'expired' ||
@@ -193,11 +83,7 @@ export function InteractiveConnect({
             <Button variant="outline" onPress={() => void openConfirmation(waiting)}>
               {t('plugins.authorization.openAgain')}
             </Button>
-            <Button
-              variant="ghost"
-              disabled={busy}
-              onPress={() => plugins.authorization.check(entry.id, method.id)}
-            >
+            <Button variant="ghost" disabled={isBusy} onPress={() => check()}>
               {t('plugins.authorization.checkAgain')}
             </Button>
           </View>
@@ -236,7 +122,7 @@ export function InteractiveConnect({
               fields={applicationFields}
               values={existingApplication.fields}
               invalidFields={existingApplication.invalid}
-              disabled={busy}
+              disabled={isBusy}
               onChange={(fieldId, value) =>
                 setExistingApplication((previous) => {
                   if (!previous) return previous;
@@ -249,27 +135,27 @@ export function InteractiveConnect({
             />
             <Button
               size="lg"
-              loading={busy}
+              loading={isBusy}
               disabled={!hasEveryField(applicationFields, existingApplication.fields)}
               onPress={() => void submitExistingApplication()}
               testID="plugin-use-existing-submit"
             >
               {t('plugins.authorization.useExistingSubmit')}
             </Button>
-            <Button variant="ghost" disabled={busy} onPress={() => setExistingApplication(null)}>
+            <Button variant="ghost" disabled={isBusy} onPress={() => setExistingApplication(null)}>
               {t('plugins.authorization.useExistingCancel')}
             </Button>
           </View>
         ) : null}
         {state && !waiting && state.status !== 'ready' && !existingApplication ? (
-          <Button size="lg" loading={busy} onPress={() => void begin()} testID="plugin-authorize">
+          <Button size="lg" loading={isBusy} onPress={() => void begin()} testID="plugin-authorize">
             {t(state.status === 'application-ready' ? `${textKey}.continue` : `${textKey}.start`)}
           </Button>
         ) : null}
         {state?.status === 'idle' && !existingApplication && applicationFields ? (
           <Button
             variant="outline"
-            disabled={busy}
+            disabled={isBusy}
             onPress={() => setExistingApplication({ fields: {}, invalid: new Set() })}
             testID="plugin-use-existing"
           >
@@ -280,7 +166,7 @@ export function InteractiveConnect({
           <ContentState.Loading title={t('plugins.authorization.finishing')} />
         ) : null}
         {error && state?.status === 'ready' ? (
-          <Button variant="outline" disabled={busy} onPress={() => void begin(true)}>
+          <Button variant="outline" disabled={isBusy} onPress={() => void begin(true)}>
             {t('plugins.authorization.reauthorize')}
           </Button>
         ) : null}
@@ -290,14 +176,13 @@ export function InteractiveConnect({
         (error || state.status === 'application-ready') ? (
           <Button
             variant="ghost"
-            disabled={busy}
+            disabled={isBusy}
             onPress={() =>
               alert.confirm({
                 title: t('plugins.authorization.resetApplication'),
                 description: t('plugins.authorization.resetApplicationMessage'),
                 confirmLabel: t('plugins.authorization.resetApplication'),
-                onConfirm: () =>
-                  void act(() => plugins.authorization.resetApplication(entry.id, method.id)),
+                onConfirm: () => void resetApplication(),
               })
             }
           >
@@ -305,10 +190,7 @@ export function InteractiveConnect({
           </Button>
         ) : null}
         {waiting || state?.status === 'ready' ? (
-          <Button
-            variant="ghost"
-            onPress={() => void act(() => plugins.authorization.cancel(entry.id, method.id))}
-          >
+          <Button variant="ghost" onPress={() => void cancel()}>
             {t('plugins.authorization.cancel')}
           </Button>
         ) : null}
